third_party/blink/web_tests/editing/selection/collapse-null-expected.txt

Ensure that collapse with null clears selection

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS selection.rangeCount is 1
PASS selection.rangeCount is 0
PASS selection.rangeCount is 1
PASS selection.rangeCount is 0
PASS successfullyParsed is true

TEST COMPLETE

hello
